Extract settings into src/core/settings.ts
Move runtime config loading, settings save handlers, setting
parsers/validators, and configurable deny reasons into a dedicated
module. Typecheck, 167 tests, and build all pass.
